backend/requirements.txt
Django
django-allauth
django-cors-headers
django-rest-auth
djangorestframework
djangorestframework-jwt
psycopg2